suppress warnings
* vm.c (rb_vm_make_proc): cast to suppress warning by VC6.

* ext/-test-/win32/console/attribute.c (console_set_attribute):
  ditto.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@48266 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/-test-/win32/console/attribute.c
@@ -25,7 +25,7 @@ console_set_attribute(VALUE io, VALUE attr)
     HANDLE h = (HANDLE)rb_w32_get_osfhandle(fd);
 
     if (h == (HANDLE)-1) rb_raise(rb_eIOError, "invalid io");
-    SetConsoleTextAttribute(h, NUM2INT(attr));
+    SetConsoleTextAttribute(h, (WORD)NUM2INT(attr));
     return Qnil;
 }
